requirements.txt
redis
fakeredis
pydantic>=2.0.0
typer
faiss-cpu
numpy
pytest
google-generativeai
pillow
pymongo
dnspython
python-dotenv
